Separate tests on the seeded portfolio landing page from tests on independent functionality
